LayoutTests/fast/viewport/viewport-134.html

<head>
    <title>Viewport META with semi-colons "width=device-width; initial-scale=2".</title>
    <meta name="viewport" content="width=device-width; initial-scale=2">
    <script>
        function test() {
            if (window.testRunner) {
                testRunner.dumpAsText();
                alert(internals.configurationForViewport(document, 1, 320, 480, 320, 352));
            }
        }
    </script>
</head>
<body onload="test();">
