LayoutTests/fast/viewport/viewport-134-expected.txt

CONSOLE MESSAGE: line 3: Viewport argument value "device-width;" for key "width" not recognized. Content ignored.
ALERT: viewport size 160x176 scale 2 with limits [2, 5] and userScalable true
